+++ CHANGES.md
@@ -1,5 +1,7 @@
 # Unreleased
 
+- Check for out of scope variables
+  [\#175](https://github.com/ocaml-gospel/ortac/pull/175)
 - Translate constant integer patterns with a guard testing for equality
   [\#174](https://github.com/ocaml-gospel/ortac/pull/174)